:truck: [cookiecutter] Move iterhooks from filestorage to main
