+++ src/cutty/filestorage/adapters/cookiecutter.py
@@ -14,17 +14,6 @@ from cutty.filestorage.adapters.disk import FileExistsPolicy
 from cutty.filestorage.adapters.disk import temporarydiskfilestorage
 from cutty.filestorage.domain.files import File
 from cutty.filestorage.domain.storage import FileStore
-from cutty.filesystems.domain.path import Path
-
-
-def iterhooks(path: Path) -> Iterator[Path]:
-    """Load hooks in a Cookiecutter template."""
-    hooks = {"pre_gen_project", "post_gen_project"}
-    hookdir = path / "hooks"
-    if hookdir.is_dir():
-        for path in hookdir.iterdir():
-            if path.is_file() and not path.name.endswith("~") and path.stem in hooks:
-                yield path
 
 
 def _runcommand(path: pathlib.Path, *, cwd: pathlib.Path) -> None:
