+++ tests/unit/application/cookiecutter/test_main.py
@@ -1,4 +1,5 @@
 """Unit tests for cutty.application.cookiecutter.main."""
+from cutty.application.cookiecutter.main import iterhooks
 from cutty.application.cookiecutter.main import iterpaths
 from cutty.filesystems.adapters.dict import DictFilesystem
 from cutty.filesystems.domain.path import Path
@@ -29,4 +30,42 @@ def test_iterpaths_other_directories() -> None:
     paths = iterpaths(path, config)
 
     assert next(paths) == path / "{{ cookiecutter.project }}"
+    assert next(paths, None) is None
+
+
+def test_iterhooks_none() -> None:
+    """It does not yield if there is no hook directory."""
+    filesystem = DictFilesystem({"{{ cookiecutter.project }}": {}})
+    path = Path(filesystem=filesystem)
+    paths = iterhooks(path)
+
+    assert next(paths, None) is None
+
+
+def test_iterhooks_paths() -> None:
+    """It returns paths to hooks."""
+    filesystem = DictFilesystem(
+        {
+            "hooks": {"pre_gen_project.py": ""},
+            "{{ cookiecutter.project }}": {},
+        }
+    )
+    path = Path(filesystem=filesystem)
+    paths = iterhooks(path)
+
+    assert next(paths) == path / "hooks" / "pre_gen_project.py"
+    assert next(paths, None) is None
+
+
+def test_iterhooks_bogus_hooks() -> None:
+    """It ignores hooks with a backup extension."""
+    filesystem = DictFilesystem(
+        {
+            "hooks": {"pre_gen_project.py~": ""},
+            "{{ cookiecutter.project }}": {},
+        }
+    )
+    path = Path(filesystem=filesystem)
+    paths = iterhooks(path)
+
     assert next(paths, None) is None
