+++ tests/unit/filestorage/adapters/test_cookiecutter.py
@@ -4,13 +4,10 @@ import pathlib
 import pytest
 
 from cutty.filestorage.adapters.cookiecutter import cookiecutterfilestorage
-from cutty.filestorage.adapters.cookiecutter import iterhooks
 from cutty.filestorage.domain.files import Executable
 from cutty.filestorage.domain.files import File
 from cutty.filestorage.domain.files import RegularFile
 from cutty.filestorage.domain.storage import FileStoreManager
-from cutty.filesystems.adapters.dict import DictFilesystem
-from cutty.filesystems.domain.path import Path
 from cutty.filesystems.domain.purepath import PurePath
 
 
@@ -121,41 +118,3 @@ def test_skip_if_file_exists(tmp_path: pathlib.Path, file: File) -> None:
         storefile(file)
 
     assert path.read_text() == "old"
-
-
-def test_iterhooks_none() -> None:
-    """It does not yield if there is no hook directory."""
-    filesystem = DictFilesystem({"{{ cookiecutter.project }}": {}})
-    path = Path(filesystem=filesystem)
-    paths = iterhooks(path)
-
-    assert next(paths, None) is None
-
-
-def test_iterhooks_paths() -> None:
-    """It returns paths to hooks."""
-    filesystem = DictFilesystem(
-        {
-            "hooks": {"pre_gen_project.py": ""},
-            "{{ cookiecutter.project }}": {},
-        }
-    )
-    path = Path(filesystem=filesystem)
-    paths = iterhooks(path)
-
-    assert next(paths) == path / "hooks" / "pre_gen_project.py"
-    assert next(paths, None) is None
-
-
-def test_iterhooks_bogus_hooks() -> None:
-    """It ignores hooks with a backup extension."""
-    filesystem = DictFilesystem(
-        {
-            "hooks": {"pre_gen_project.py~": ""},
-            "{{ cookiecutter.project }}": {},
-        }
-    )
-    path = Path(filesystem=filesystem)
-    paths = iterhooks(path)
-
-    assert next(paths, None) is None
